video

+++ app/views/content/front/profile/video.blade.php
@@ -61,7 +61,7 @@
 							<source src="/{{ $item->path }}" type="video/mp4">
 							Ваш браузер не поддерживает HTML5 видео.
 						</video>
-						@if($userInfo->user_id == Auth::user()->id)
+						@if(Auth::check() && $userInfo->user_id == Auth::user()->id)
 							<a href="/video/delete/{{ $item->id }}" class="fa fa-times delete-image" onclick="return confirm('Удалить?')?true:false;" style="top:5px;right:0px;"></a>
 						@endif	
 					</div>
